fixed some suggestions from codepro analytix

+++ trunk/BaseletElements/src/com/umlet/element/experimental/settings/SettingsUseCase.java
@@ -13,7 +13,7 @@ public class SettingsUseCase extends Settings {
 	public XValues getXValues(double y, int height, int width) {
 		double b = height/2.0f;
 		double a = width/2.0f;
-		double x = (double) Math.sqrt((1-(Math.pow(b-y, 2) / Math.pow(b, 2)))*Math.pow(a, 2));
+		double x = Math.sqrt((1-(Math.pow(b-y, 2) / Math.pow(b, 2)))*Math.pow(a, 2));
 		return new XValues(a-x, a+x);
 	}
